requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attributedict==0.3.0
attrs==23.1.0
auto-gptq @ https://github.com/jllllll/AutoGPTQ/releases/download/v0.5.1/auto_gptq-0.5.1+cu121-cp310-cp310-win_amd64.whl#sha256=d22bb6d1f7f22e3c78fd8bf36ce23ec458ef44453abf28311c7715896d1f28bc
autoawq==0.1.7
awscli==1.32.61
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
blessings==1.7
blinker==1.9.0
botocore==1.34.61
build==1.2.2.post1
cachetools==5.3.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
Chroma==0.2.0
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.2.2
comtypes==1.4.8
contourpy==1.2.0
coverage==7.3.2
cramjam==2.7.0
ctransformers @ https://github.com/jllllll/ctransformers-cuBLAS-wheels/releases/download/AVX2/ctransformers-0.2.27+cu121-py3-none-any.whl#sha256=9be6bfa8ac9feb5b2d4c98fbf5ac90394bbfa5c406313f8161dca67b28333e51
cycler==0.12.1
dataclasses-json==0.6.7
DataProperty==1.0.1
datasets==2.15.0
debugpy==1.8.7
decorator==5.1.1
deepdiff==6.7.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distlib==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
docutils==0.16
durationpy==0.9
einops==0.7.0
exceptiongroup==1.1.3
executing==2.1.0
exllama @ https://github.com/jllllll/exllama/releases/download/0.0.18/exllama-0.0.18+cu121-cp310-cp310-win_amd64.whl#sha256=cf248071ae070a4b229cb3daff17f06d28e2b0cf990fcc9b851b67df407da354
exllamav2 @ https://github.com/turboderp/exllamav2/releases/download/v0.0.8/exllamav2-0.0.8+cu121-cp310-cp310-win_amd64.whl#sha256=0e6046d5377139562c2ccab4b1a3deffa1ecb46490e0b22bbd1abe9d1f7bcced
faiss-cpu==1.9.0.post1
fastapi==0.104.1
fastparquet==2023.10.1
ffmpy==0.3.1
filelock==3.16.1
flash-attn @ https://github.com/bdashore3/flash-attention/releases/download/2.3.2-2/flash_attn-2.3.2+cu122-cp310-cp310-win_amd64.whl#sha256=bc1feca4c9888d820d64227d36cc5100dd7db3c78514e7753b6dda3c3e25dd5c
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2024.10.0
gekko==1.0.6
gemini-api==0.1.6
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.152.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-texttospeech==2.21.1
google-cloud-vision==3.8.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gptq-for-llama @ https://github.com/jllllll/GPTQ-for-LLaMa-CUDA/releases/download/0.1.1/gptq_for_llama-0.1.1+cu121-cp310-cp310-win_amd64.whl#sha256=cbb115bd5f493173756826d5fe39293e1ab7bad818fde43886f9cf7722d10d71
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.1.1
groq==0.12.0
grpcio==1.67.1
grpcio-status==1.67.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib-resources==6.1.1
importlib_metadata==8.4.0
inspecta==0.1.3
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.5
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.144
llama_cpp_python @ https://github.com/abetlen/llama-cpp-python/releases/download/v0.2.18/llama_cpp_python-0.2.18-cp310-cp310-win_amd64.whl#sha256=e06f7341b8876c7e798276435a5435fa7f34c8c85b3843ac69daee4df0b03ba2
llama_cpp_python_cuda @ https://github.com/jllllll/llama-cpp-python-cuBLAS-wheels/releases/download/textgen-webui/llama_cpp_python_cuda-0.2.18+cu121-cp310-cp310-win_amd64.whl#sha256=425b14752142025908fd4e59e16dc90c26a8978813b81fd76cbee8ddc9e10363
lm-eval==0.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.25.0
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.3.3
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
optimum==1.14.0
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.0
peft==0.6.2
pillow==10.4.0
platformdirs==3.11.0
pluggy==1.3.0
portalocker==2.8.2
posthog==3.7.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.6
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyarrow-hotfix==0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycountry==22.3.5
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.17.1
PyMySQL==1.1.1
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
pypiwin32==223
pyproject-api==1.6.1
pyproject_hooks==1.2.0
pyreadline3==3.4.1
pytablewriter==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pyttsx3==2.98
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.31.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.2
rootpath==0.1.1
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.13.0
rsa==4.7.2
s3transfer==0.10.0
sacrebleu==1.5.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.1.99
sentry-sdk==1.35.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.39.0
sympy==1.13.1
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
tesseract==0.1.3
texttable==1.7.0
threadpoolctl==3.5.0
tokenizers==0.20.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1+cu117
torch-grammar @ git+https://github.com/oobabooga/torch-grammar.git@82850b5383a629f3b0fa1fba7d8f2aba3185ddb2
torchaudio==2.1.1
torchvision==0.15.2+cu117
tornado==6.4.1
tox==4.11.3
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.46.3
typepy==1.3.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.24.0.post1
virtualenv==20.24.6
wandb==0.16.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.20.2
zstandard==0.22.0
